chore: Simplify state machine implementation of prosemirror prompt input (#7154)
In this PR I'm reducing the API surface of the state machine and move
most of the events related to
value and mentions changes out of it. Hopefully this makes the state
machine easier to understand
and to work with.

Now the state machine exposes an `document.update` event whose payload
is a function that gets passed
the current editor state and returns a transaction. Note however that
these changes actions are still
tested toghether with the state machine in `promptInput.test.ts`. I
think it's important to have one
place to test all the expected behavior of the prompt input.

The main reason behind this change is that the current implementation is
very much driven by the current
API the `PromptInput` component exposes. But going forward we might
change/reduce/expand this API.
Ideally the state machine won't have to be changed for that.

In an upcoming PR I plan to refactor the mentions menu implementation to
be more generic to better
support menu items that are not `ContextItem`s.

## Test plan

Unit test