Improve query rewrite for search context (#7164)
When gathering the initial search context for a repo, we use the 'fast'
LLM to create a keyword search from the user's question. In evals, we
found that this query rewrite made things worse for simple questions, so
[we only enabled it for longer or non-English
questions.](https://github.com/sourcegraph/cody/pull/4494). I recently
dug into this, and figured out this was because the rewrite prompt was
not performing well.

This PR improves the rewrite prompt so we can remove the special-casing
around short English questions. Benefits:
* Removes a "gotcha" that was making the mental model more complicated.
* It will let us switch to [NLS search instead to power context
search](https://linear.app/sourcegraph/issue/SPLF-852/use-search-api-for-cody-context),
which always expects keyword-style queries.
* The improved query rewrite is significantly faster. Here is the
average rewrite time during the "Tool Call" eval:
    * Before: `1556.10ms`
    * After: `677.25ms`

Downside: single-sentence Chat questions with an @-repo mention become a
bit slower, as they now rewrite the query.

**FAQ: why are you working on this? We're reworking all of Chat!**
This didn't take me much time at all. Also I need this to unblock my PR
to move to Cody context to NLS search under-the-hood and eliminate the
confusing duplication.

Relates to SPLF-852