refactor(telemetry): opt-out option (#7205)
This commit update the ability to disable telemetry for Enterprise-users
only and no longer works for PLG users to ensure we can improve the
product.

- Updated the `cody.telemetry.level` configuration to reflect that
telemetry can only be disabled for Enterprise instances.
- Updated the telemetry provider to disable telemetry if the user is NOT
on sourcegraph.com and has telemetry disabled.

This change ensures that telemetry is only disabled for Enterprise
users, as telemetry is always enabled for sourcegraph.com users.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Makes sure the description for the telemetry level configuration shows
that it is for enterprise only:


![image](https://github.com/user-attachments/assets/0f72a984-0563-4355-8c01-efb02a198df1)

Verify telemetry is not disabled even if `cody.telemetry.level` is set
to off.

---------

Co-authored-by: Emi <emi@hexops.com>