feat(smart-apply): prefetch everything (#6768)
- [Slack post with demo
recordings](https://sourcegraph.slack.com/archives/C89KCDK5J/p1737685247830649).
- For each code block in the chat window, it prefetches the smart apply
selection and the relevant edit command result.
- It caches the selection and edit command promises, so if a user clicks
on the smart apply button before the prefetch process is done, we can
reuse them. The edit command provider is updated to apply part of the
streamed changes right away if we already have them in the cache, and it
switches to streaming mode, which is shown in the UI.
- The relevant code paths are cleaned up a bit to allow for a modular
fetching approach for the smart apply edit command. Initially, it was
implemented as an atomic operation that was hard to split into separate,
retrievable stages. This current implementation isn't perfect, but it
improves things a bit, and it's a good starting point for further
improvements.
- Closes
https://linear.app/sourcegraph/issue/CODY-5006/instant-smart-apply-with-preloading
- [Cost estimates in
Slack](https://sourcegraph.slack.com/archives/C08DP4R5CG0/p1739944377531849?thread_ts=1739796101.434469&cid=C08DP4R5CG0).
-[Added](https://github.com/sourcegraph/cody/pull/6768/commits/288929cee9009fff7305644180ed98f16741581f)
a feature to ensure this feature is disabled by default. You can test it
locally by adding an override for your user on DotCom