fix: Make @-button work in prompt editor v2 (#7195)
This PR makes the @-button in the editor toolbar work with the new
prompt input.

To do this I added a new method to the editor API, `openAtMentionMenu`,
instead of just appending `@`.

Note that unlike the current input we don't have to check whether there
is already an `@` at the end of the input
because clicking the button will already 'deactivate' any at-mention
that is currently present and we cannot 'reuse'
at-mentions.

Closes
https://linear.app/sourcegraph/issue/SRCH-1692/prosemirror-clicking-the-button-doesnt-show-the-mentions-menu

## Test plan

Manual testing. Clicking the @-button with the new input enabled inserts
an at-mention and focuses the editor.
The existing input continues to work as expected.