feat(user-menu): Add billing metadata to telemetry events
This commit adds billing metadata to the telemetry events for the user menu and the upgrade to pro link. This will allow us to track the usage of these features and attribute them to the correct product and category.

This change is part of the effort to improve the telemetry data for Cody features.