feat(commit generation): prioritize Gemini Flash models  (#7252)
This commit updates the Cody Source Control to prioritize Gemini Flash
models when selecting a model for edit commands. It filters out
experimental models and then selects Gemini Flash Lite or Gemini Flash
if available, falling back to the first available model if neither is
found.

This change ensures that the edit commands utilize the most performant
and suitable models.


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

- Verify that the edit commands use Gemini Flash models if available.
- Verify that the edit commands fall back to other models if Gemini
Flash is not available.

Connect to demo.sourcegraph and verify this change works on older
instances