fix: Fix at-mentions menu in prompt template
Without calling createNewChat the at-mentions menu is not properly
populuated in prompt templates. That's because we don't have a valid
panel ID.
Calling createNewChat solves this but feels like the wrong solution
since we do not actually want to create a new chat.
I think it's fine as a stop-gap solution, since that's what the previous
code this as well, but we should find a more reasonable solution.