fix(dev): Refine image upload logic and vision model support (#7305)
Follow up on https://github.com/sourcegraph/cody/pull/7279

This commit refines the logic for enabling image uploads and managing
vision model support in the Cody application.

Key changes:

- Updated the `isImageUploadEnabled` function in `Toolbar.tsx` to
correctly determine when image uploads are allowed. Image upload is now
enabled if the user is not on Sourcegraph.com OR is using a BYOK model,
AND the selected model has the `Vision` tag.
- Modified the `syncModels` function in `lib/shared/src/models/sync.ts`
to conditionally remove the `Vision` tag from models provided by Google
when running on Enterprise instances with early access enabled. This is
because the gateway does not yet support vision models for Google.


## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->


Follow instruction in https://github.com/sourcegraph/cody/pull/7279