change(commands): use latest sonnet for fix and latest haiku-3-5 for doc intent (#7241)
Refs CODY-5114
Latest sonnet was suggested and haiku 3 is deprecated as such try to
document
with flash 2.0 since it's comparable in speed and cost to the old
haiku-3.

EDIT:
Upon testing flash-2.0, haiku-3, gpt-4o-mini and haiku-3.5 for docs the
best model has been haiku-3.5 at following the prompt.

There are still some commands where we use the older sonnet, but that
can be
changed in another pull request.


## Test plan
Use the fix or doc intent and see if the operations are performed
correctly.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Tom Ross <tom@umpox.com>