chore(auto-edit): capture request/response metadata from auto-edit API adapters (#7302)
- Bubbles up request/response metadata from auto-edit API adapters to
the autoedit provider so we can later propagate this data to the
auto-edit debug panel. It would allow us to quickly inspect the network
traffic from the debug panel and know precisely what goes in and out
without attaching the debugger to the extension.
- Only Cody Gateway and Sourcegraph completions providers are updated to
fully support newly added optional metadata properties since we use them
the most (enterprise and PLG setups). We can update the rest of the
adapters if needed in a follow-up.