fix(webview): show copy button in web clients
This commit refactors the UI for code blocks in chat messages to improve the placement of action buttons (copy, apply, insert) and fixes the issue where copy button is not showing up on web.

The cause of the issue is due to the logic where actionsContainer will only be displayed when previewContainer is available, a regression from https://github.com/sourcegraph/cody/pull/7217.

The fix is to always append available actionsContainer, but also insert the previewContainer before that when available.

Key changes:

- Modified CSS styles in `ChatMessageContent.module.css` to adjust padding, borders, and background colors for code blocks and buttons.
- Updated `ChatMessageContent.tsx` to insert the preview container (containing file info) before the code block and the actions container after the code block.
- Updated `create-buttons.ts` to use the default `buttonsContainer` style instead of `buttonsContainerTransparent` for both preview and action button containers.

## Test Plan

Added storybook to cover this component