feat(auto-edit): full debug panel UI implementation (#7306)
- Implements UI for the auto-edit debug panel based on the data we
already have in store. I'll share a demo video in Slack later today.
- Built on top of https://github.com/sourcegraph/cody/pull/7304
- There are several minor issues with the UI. E.g., context summary
shows the total number of items as 0 while it's definitely not. The
side-by-side diff view highlighting is inaccurate for some inline
changes. We address such problems in follow-ups or keep them as is
because this UI is for internal use, so we should not spend much extra
time on it.