fix(completions): handle Cloudflare 403 errors with VPN message
When Cloudflare blocks a request (status code 403) due to a VPN or untrusted network, provide a more informative error message to the user. The message will instruct the user to try again without a VPN and includes the Cloudflare Ray ID for debugging.
We do this without waiting for the response to complete, as the error is already known.