fix(autoedit): Fix conflict between suggestions and completion menu (#7389)
https://github.com/user-attachments/assets/5d105794-516b-44dd-b81f-b0c991c835a1



## Description

Fixes an issue where we attempt to suggest an auto-edit that could
conflict with the completion menu in a bad way. E.g. it may delete or
add code that doesn't reflect the selected item in the menu. If the user
was to accept this item with Tab, the edit would be accepted including
the unwanted code.

## Test plan

- Unit tests
- Manual testing with completion items