feat(auto-edit): cancel redundant requests (#7400)
- Propagates the abort signal to the fetch call used for auto-edit
requests, ensuring that we cancel them based on the logic in the
auto-edit provider. This should reduce the load on our inference
deployment.
- Instead of throwing an error on cancellation, we return an object
indicating to the auto-edit provider that the request was cancelled.
This allows us to maintain the same data flow for the integration with
the analytics logger and the debug panel.