feat(auto-edit): reduce context size (#7401)
- Reduces the context limits:
    - Document prefix and suffix from 5000 to 1000. 
    - Recent edits from 1500 to 2500
    - Diagnostics retriever from 500 to 250
- No quality evaluations are available for this change yet. We will work
on it with @hitesh-1997 in the next couple of weeks. This change's goal
is to reduce latency and is based on the idea that the document prefix
and suffix are less important than recent edits. Since the model is
relatively small, it likely does not benefit from a long context window.
For example, in autocomplete, our total context limit is 2k tokens. The
diagnostics retriever limit is reduced based on manual testing, seeing
that most of the errors are not actionable for the model, and usually,
having a couple of them is enough.