fix(models): Pass models to getModelByID to allow for custom model lists
This commit updates the `getContextWindowByID` function to pass the `models` array to the `getModelByID` function. This change allows for the use of a custom model list when retrieving context window information. This is useful in scenarios where a different set of models needs to be considered than the default set.