chore(auto-edit): fix accept/reject analytics events (#7427)
- @umpox, when an inline completion is accepted, the
`onDocumentSelectionChange` callback is triggered before the acceptance
callback, which means we were marking suggestions as rejected right
after they were accepted and right before the acceptance callback was
triggered. For some reason, the order of the callbacks is different in
E2E tests (I tried to capture this bug there). But you can verify the
unexpected behavior locally in the extension host with the auto-edit
debug panel by generating suggestions with inline completions only and
accepting them.
- Adds reject reason fields to the analytics events.