fix(completions): Add api-version check for completions endpoint (#7447)
The `api-version` parameter is only supported for API version 1 and
above. This commit adds a check to ensure that the `api-version`
parameter is only appended to the URL if the API version is greater than
or equal to 1. This prevents potential errors when using older or
invalid API versions.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

See updated recordings

(cherry picked from commit a4287b9f91b2b33bbd5f029fb701ee222bcd9807)