feat(auto-edit): Add optimized prompt structure to leverage the prompt cache (#7453)
As part of the prefil optimization, add the prompt structure to leverage
the maximum KV cache.
The new prompt provider add the prompt in the following order:
- Other files prompt components are added to the top. Since most of the
edits happen in the current file, most of the prompt from the other
files (`view-port` and `recent-edits`) are added to the top.
- For the most recent added `recent-edits` it sets a minimum time. They
are still added in the usual location i.e. near the end of the file
since they are important.
- Adds a cap on the number of snippets for the `recent-view-port`. Check
the [details
here](https://sourcegraph.slack.com/archives/C08GUBSHMKN/p1742306288375029)
for more details.

## Test plan
- Added unit tests