Prevent parallel agent restarts from happening (#7466)
Fixes
https://github.com/sourcegraph/cody/issues/7446#issuecomment-2730388728

Potential fix for the issue with multiple agents spawned in parallel.
In our current restart code we do not sufficiently guard us against
parallel cody restarts.

In what situations it can happen?
For example if agent is starting doing multiple independent actions
(e.g. syncing some files, updating chat state, etc) and it will crash,
each of those action can attempt to restart a server. That should be
limited to actions explicitly triggered by user, but I also fixed one
place where it was not (and which could have been source of the
problem).
Regardless, we should also safeguard it on the level of the `restart`
function.

## Test plan

1. Run 'Cody: Restart Cody' a dozen of times as quickly as you can (you
can assign it shortcut to make it easier)
2. Using process monitor make sure number of the agent processes is not
increasing
3. Make sure cody is functioning properly after restart is done