feat(agentic chat): add internal feature flag (#7479)
This commit introduces an internal feature flag,
`NextAgenticChatInternal`, to control the availability of the new
agentic chat experience. This flag is intended for internal use only and
will not be exposed to users.

The changes include:

- Adding the `NextAgenticChatInternal` feature flag to the `FeatureFlag`
enum.
- Updating the `ChatController` to use the new feature flag to enable
the experimental agentic chat features.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

Simple change. Check if feature flag is correct