feat(vscode): prioritize 'flash' model over 'flash-lite' (#7386)
Closes CODY-5281. This PR depends on
https://github.com/sourcegraph/cody/pull/7385.

This commit changes the order in which the `flash` and `flash-lite`
models are assigned to `this.model` in `CodySourceControl`.

The `flash` model is now prioritized over the `flash-lite` model. This
ensures that the more capable `flash` model is used when available,
improving the overall performance and accuracy of Cody's code editing
features.

The comment was also updated to use the word "Remove" instead of
"Removes" to align with the code.

## Test plan

Manually test by comparing the performance and quality for both.