fix(auto-edit): Remove unnecessary prompt whitespace (#7481)
## Description

Our prompt strategy usually consists of bulding `PromptStrings`
together. When we don't have valid information to prompt we return an
empty string. I think this is OK and a good way to build the prompts,
but we had some cases where we were unnecessarily padding empty strings
with new lines. This fixes that.

## Test plan

Tested with unit tests, see the updated snapshots

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->