fix(vscode): Ensure file link clicks work in ToolStatusCell
Close https://linear.app/sourcegraph/issue/CODY-5444/bug-search-results-and-file-tool-results-are-not-clickable

The Issue: In the original implementation, if the vscodeAPI prop was undefined during initial component render (because the VSCode API wasn't fully initialized yet), the onFileLinkClicked function would be created with a reference to that undefined value. Even if the API became available later, the callback would still reference the initial undefined value due to how closures work.

The Fix: By calling getVSCodeAPI() at the time the link is clicked (rather than capturing the value during render), the code now gets the most current API reference at the exact moment it's needed. This guarantees that even if the API wasn't available during the component's first render, the link functionality will still work when clicked as long as the API is available by then.