feat(auto-edit): Handle accept events correctly when document changes match the active request (#7483)
## Description

Notable changes in this PR:
1. We now mark auto-edits as accepted when the document changes match
the suggestion. I think this is logically correct for our telemetry - if
a user pastes some code in that matches our suggestion we can say that
our suggestion was accepted and pat ourselves on the back
2. We handle a nasty bug where VS Code makes an insertion for native
completions _before_ calling the command we provide
`cody.supersuggest.accept`. This ordering means that, due to how we try
to intercept cursor changes as a way to reject suggestions, we would
incorrectly reject an accepted completion. See the flow below:

### Completion acceptance bug
Before:
1. Suggest native completion
2. User accepts it
3. VS Code inserts the text, and calls `onDidChangeTextDocument` and
`onDidChangeTextEditorSelection` due to the inserted text and cursor
position change.
4. Our logic intercepts `onDidChangeTextEditorSelection` and sees that
the user moves their cursor, we reject the suggestion
5. VS Code calls `cody.supersuggest.accept`, we cannot accept the change
in our telemetry as we have already rejected it.

After:
1. Suggest native completion
2. User accepts it
3. VS Code inserts the text, and calls `onDidChangeTextDocument` and
`onDidChangeTextEditorSelection` due to the inserted text and cursor
position change.
4. Our new logic intercepts `onDidChangeTextDocument`, sees that the
insertion matches our active request and marks the suggestion as
accepted

## Test plan
- [x] Verify auto-edits work correctly when accepting via different
methods
- [x] Check that analytics capture the correct acceptance reasons