feat(autoedit): Enable Agent support (#7311)
## Description

This PR enables support for autoedit for clients other than VS Code.

### Implementation

This is an extension of the existing `autocomplete/execute` call that
clients make. The idea is that this is a non-breaking change that
becomes an enhancement for existing autocomplete.
 
We have deprecated the `items` property and now have two explicit fields
`inlineCompletionItems` and `decoratedEditItems`. These cover the two
types of an autocomplete: a completion and an edit.

@mkondratek Is working on integrating this API into JetBrains, so we may
make some changes but this has already been confirmed to be working for
both JetBrains and the fake clients I have created in my integration
tests.


## Test plan

1. Integration tests via the Agent API with a range of
clientCapabilities
2. Actively testing this in a real client (JetBrains). This work is also
feature flagged and behind a setting.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->