feat: Add Dockerfile that installs cody from npm
This is just so that we have a reference here for now. We're unblocking
a customer while we integrate a CI  pipeline that will publish this
whenever an agent release goes out. That means coping the _built_ files
into this container, rather than using npm. We'll also need to give the
GHA the creds to publish this to Dockerhub.

I have chosen not to publish this image given it doesn't currently exist
in the Sourcegraph Dockerhub registry. If the customer demands it, I
could.